requirements.txt
absl-py==0.7.1
astor==0.7.1
beautifulsoup4==4.6.3
bert-serving-client==1.8.9
bert-serving-server==1.8.9
boto==2.49.0
boto3==1.9.83
botocore==1.12.83
bz2file==0.98
certifi==2018.11.29
chardet==3.0.4
docutils==0.14
enum34==1.1.6
gast==0.2.2
gensim==3.7.0
GPUtil==1.4.0
grpcio==1.19.0
h5py==2.9.0
idna==2.7
imbalanced-learn==0.4.3
jmespath==0.9.3
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.1
mock==2.0.0
nltk==3.4
numpy==1.16.2
pandas==0.24.2
pandas-ml==0.6.1
pbr==5.1.3
pkg-resources==0.0.0
protobuf==3.7.1
pyahocorasick==1.3.0
python-dateutil==2.7.5
pytz==2018.7
PyYAML==5.1
pyzmq==18.0.1
requests==2.20.1
s3transfer==0.1.13
scikit-learn==0.20.1
scipy==1.2.0
singledispatch==3.4.0.3
six==1.11.0
smart-open==1.8.0
soupsieve==1.6.1
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
tqdm==4.31.1
urllib3==1.24.1
Werkzeug==0.15.2
